Header updates, modified tamtypes to more accurately reflect the word sizes

+++ pspsdk/psp/include/tamtypes.h
@@ -14,28 +14,28 @@
 #ifndef _TAMTYPES_H_
 #define _TAMTYPES_H_ 1
 
-typedef	unsigned char 		u8;
-typedef unsigned short 		u16;
+typedef	unsigned char 			u8;
+typedef unsigned short 			u16;
 
 typedef	volatile unsigned char 		vu8;
 typedef volatile unsigned short 	vu16;
 
-typedef unsigned long int	u32;
-typedef unsigned long long	u64;
+typedef unsigned int			u32;
+typedef unsigned long long		u64;
 
-typedef volatile unsigned long int	vu32;
+typedef volatile unsigned int		vu32;
 typedef volatile unsigned long long	vu64;
 
-typedef signed char 		s8;
-typedef signed short 		s16;
+typedef signed char 			s8;
+typedef signed short 			s16;
 
-typedef volatile signed char	vs8;
-typedef volatile signed short	vs16;
+typedef volatile signed char		vs8;
+typedef volatile signed short		vs16;
 
-typedef signed long int		s32;
-typedef signed long long	s64;
+typedef signed int			s32;
+typedef signed long long		s64;
 
-typedef volatile signed long int	vs32;
+typedef volatile signed int		vs32;
 typedef volatile signed long long	vs64;
 
 #ifndef NULL
